refactor(appearance): split handlers/appearance.rs into a directory
The single-file handler hit 429 lines, over the 400-line soft limit
documented in docs/rust-module-conventions.md. Split into a directory
along clean domain boundaries (matches the established
handlers/notes/ pattern):

- mod.rs              — module declarations + explicit re-exports
- paths.rs            — `.aura/` directory resolution + filename consts
- metadata.rs         — JSON file read/write (appearance.json)
- image_asset.rs      — shared PUT/GET/DELETE for image assets, magic
                        bytes, atomic writes, format-swap cleanup,
                        plus the detect_image_format unit tests
- banner.rs           — banner-specific endpoints wrapping image_asset
- background_image.rs — background-image endpoints, same wrapping

`mod.rs` only declares + re-exports, per the convention. The
ImageAssetSpec consts (BANNER_ASSET, BACKGROUND_ASSET) live with their
respective specific-asset modules.

No behavior change. Tests still pass (4/4 detect_image_format cases).
Router import path unchanged — `crate::handlers::appearance::*`
resolves to mod.rs the same way it resolved to appearance.rs before.
